abstracted behaviours into external file, tested build
