+++ src/lib/TEISimple.svelte
@@ -5,39 +5,14 @@
     import CETEI from 'CETEIcean';
     import { ProgressRadial } from '@skeletonlabs/skeleton';
     import { base } from "$app/paths";
+    import {teiBehaviours} from '../utils/teiBehaviours';
 
     let loaded = false;
     export let path = '';
 
     onMount(async () => {
         var cetei = new CETEI();
-        cetei.addBehaviors({ "tei": {
-            "note": function(elt){
-                if (!this.noteIndex){
-                    this["noteIndex"] = 1;
-                } else {
-                    this.noteIndex++;
-                }
-                let id = "note" + this.noteIndex;
-                let link = document.createElement("a");
-                link.setAttribute("id", "src" + id);
-                link.setAttribute("href", "#" + id);
-                link.innerHTML = this.noteIndex;
-                let content = document.createElement("sup");
-                content.appendChild(link);
-                let notes = this.dom.querySelector("ol.notes");
-                if (!notes) {
-                    notes = document.createElement("ol");
-                    notes.setAttribute("class", "notes");
-                    this.dom.appendChild(notes);
-                }
-                let note = document.createElement("li");
-                note.id = id;
-                note.innerHTML = "<a href=\"#src" + id + "\">^</a> " + elt.innerHTML
-                notes.appendChild(note);
-                return content;
-                }
-        }});
+        cetei.addBehaviors(teiBehaviours);
         cetei.getHTML5(path, function(data) {
             document.getElementById('TEI-container').appendChild(data);
         });
